third_party/WebKit/LayoutTests/inspector/console/console-object-constructor-name-expected.txt

CONSOLE MESSAGE: line 12: [object Object]
CONSOLE MESSAGE: line 13: [object Object]
CONSOLE MESSAGE: line 14: [object Object]
Tests that the name of the function invoked as object constructor will be displayed as its type in the front-end. Bug 50063.

console-object-constructor-name.html:12 Parent
console-object-constructor-name.html:13 Child
console-object-constructor-name.html:14 inner
